+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog pwcli
 
+## [v2.9.0 - 2024-05-24]
+### New
+- add scram hash method
+### Changed
+- add prefix option to hash command 
+- update dependencies
+
 ## [v2.8.0 - 2024-03-17]
 ### New
 - add method kms get,list and encrypt for Amazon KMS Service